requirements.txt
langchain==0.3.25
langchain-openai==0.3.17
langchain-community>=0.3.24
pydantic>=2.0.0
python-dotenv>=1.0.0
aioconsole>=0.6.2
scikit-learn==1.6.1
langchain-core>=0.1.0
numpy>=1.24.0
phrasemachine==1.0.7
umap-learn>=0.5.5
joblib>=1.3.0
spacy>=3.7.0
nltk>=3.8.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl